+++ test/test_helper.exs
@@ -1,2 +1,1 @@
 ExUnit.start(exclude: [:integration])
-Code.require_file("support/transport_mock.ex", __DIR__)